Declarative Changes
Adjusted list views on customer support cases to display the Account Name and not Contact Name. Added Parking Spot field on Serial Number to assist the Sales team with locating delivered untagged boats in the lot. Added a Due Date field on Payable that doesn't get offset by the Vendors preference for the new AP process.  Adjusted Submit for Payment flow to accommodate this new field.