Modified Off-season Survey Process
The existing process would send off a specific boat show survey when an opportunity with a boat show campaign associated would reach BOM approval pending. Because of COVID, we are now sending a retail experience survey during the off-season regardless of the campaign associated to the opportunity when the deal reaches BOM approval pending.