closes SFISSUES-470
Boat Specialists will now receive a task to call the customer to coordinate for early pickup of the retail ERP is updated to Pending Final Rigging and the Delivery Date is >= 2 months from then. Boat Specialist will now also receive a task 10 days after the delivery date of a 'Delivered' retail ERP to follow up with the customer and see how their experience was.