Restructured ERPOrder on creation flow
Deleted 'ERP Order: Notify Sales Manager when Retail ERP is created' flow and created a more generic ERP Order flow that runs only on record creation. Rather than just handling notifications for the sales manager, this flow will handle all notifications that need to happen on ERP creation regardless of the record type. This is to circumvent the limitation with the Retail ERP Notification flow not triggering on default stage.