Modified PO Approval Process entry criteria
Changed the entry criteria for the Whitefish,Barrie,Montreal and Wholesale approval process. It will now make reference to the Warehouse field on the PO rather than the users Warehouse. This allows the user to change the warehouse on the PO before submitting for approval in the event a Wholesale user creates a PO for Whitefish. This happened with Kayleigh Santi today. Had to create a new formula field as Approval Processes cannot make reference to custom lookup fields. REALLY SALESFORCE?